+++ common_AOK/usr_local_bin/network-check.sh
@@ -12,10 +12,14 @@
 if ping -c 2 8.8.8.8 >/dev/null 2>&1; then
     if ping -c 2 amazon.com >/dev/null 2>&1; then
         echo "Connected to the Internet and DNS is resolving!"
+	ex_code=0
     else
         echo "***  DNS does not seem to resolve!"
+	ex_code=2
     fi
 else
     echo "***  Not able to access the Internet!"
+    ex_code=1
 fi
 echo
+exit "$ex_code"